up-to-date library & objectalgebras

+++ Java_ObjectAlgebras/ObjectAlgebras/src/trees/SybAlg.java
@@ -4,6 +4,7 @@ import java.util.List;
 
 import com.zewei.annotation.processor.Algebra;
 
+//BEGIN_SYB_TREE
 @Algebra
 public interface SybAlg<Company, Dept, SubUnit, Employee, Person, Salary>{
   public Company C(List<Dept> depts);
@@ -12,5 +13,6 @@ public interface SybAlg<Company, Dept, SubUnit, Employee, Person, Salary>{
   public SubUnit DU(Dept dept);
   public Employee E(Person p, Salary s);
   public Person P(String name, String address);
-  public Salary S(double salary);
+  public Salary S(float salary);
 }
+//END_SYB_TREE
